fixes after CR + lib version bump
