+++ package.json
@@ -46,7 +46,7 @@
     "@next/mdx": "12.0.4",
     "@oasisdex/addresses": "0.0.8",
     "@oasisdex/automation": "1.4.0",
-    "@oasisdex/dma-library": "0.3.4",
+    "@oasisdex/dma-library": "0.3.6",
     "@oasisdex/multiply": "^0.2.11",
     "@oasisdex/oasis-actions": "0.2.16",
     "@oasisdex/transactions": "0.1.4-alpha.0",
